parser/requirements.txt
Flask
pandas
google-cloud-bigquery
coverage
pytest
pyarrow>=10.0.0